Sorta works...? (Update below)
Hi everyone,
here is the update:

Though each mole/button pair works individually when the timers are hard-coded, they are faulty when using the random timers (Also, its now using the built in randomizer instead of the external module since it was having trouble working with it).

What works:

- Multi-mole testbench has the proper values when manually shortining the mole_timers (commented out code), both for the individual and ramdomized versions.

- Dynamic hammer timer shortening should work, but Vivado decided to be sad as I was going to test it (detailed below)

Additional comment: before the code was running smoothly when lives and score both were 4 bit values. However, to represent the score of 100 and due to the limit of LEDs on the board, I shifted it so lives is now a 2 bit value and score is now 7bit. This required some minor changes in the constraint file, whack_a_mole_advanced, and Debounced_WAM_advanced modules. Unfortunatly, Vivado dicided to act up as I was doing this change so I was unable to verify if this changed worked (detailed below), so if synthesis failed try going back to the 4x4




Current issues:

- Vivado is being dumb and is saying the final case has bad syntax when it does not...I tried resetting and reloading but the issue seems to persist on my computer and It cannot synthesize without it being fixed. It systhesized earlier with the same line just fine, so not sure what the issue is here.

- Running this code with the randomized timers causes the code to stall before any mole appears on the board. To me, this would imply an error in the logic around the if statment around line 180, however the code works fine when doing one timer at a time so again, not sure whats going bad here.

- Posibly related to Vivado being dumb (not sure); game logic fails to transition from END_SCREEN state (the one Vivado is glitching on) to the IDLE state upon a button press. This issue also exists in the normal whack-a-mole code aswell.